chore(zql): refactor some test helpers
