+++ packages/zqlite/src/test/source-factory.ts
@@ -4,8 +4,12 @@ import type {Source} from '../../../zql/src/ivm/source.js';
 import type {SourceFactory} from '../../../zql/src/ivm/test/source-factory.js';
 import {createSilentLogContext} from '../../../shared/src/logging-test-utils.js';
 import {compile, sql} from '../internal/sql.js';
-import {TableSource} from '../table-source.js';
+import {TableSource, toSQLiteTypeName} from '../table-source.js';
 import type {PrimaryKey} from '../../../zero-protocol/src/primary-key.js';
+import type {QueryDelegate} from '../../../zql/src/query/query-impl.js';
+import {normalizeTableSchema} from '../../../zero-schema/src/normalize-table-schema.js';
+import type {Schema} from '../../../zero-schema/src/schema.js';
+import {MemoryStorage} from '../../../zql/src/ivm/memory-storage.js';
 
 export const createSource: SourceFactory = (
   tableName: string,
@@ -26,3 +30,50 @@ export const createSource: SourceFactory = (
   db.exec(query);
   return new TableSource('zqlite-test', db, tableName, columns, primaryKey);
 };
+
+export function createQueryDelegate(fullSchema: Schema): QueryDelegate {
+  const db = new Database(createSilentLogContext(), ':memory:');
+  const sources = new Map<string, Source>();
+  return {
+    getSource: (name: string) => {
+      let source = sources.get(name);
+      if (source) {
+        return source;
+      }
+      const schema = normalizeTableSchema(fullSchema.tables[name]);
+
+      // create the SQLite table
+      db.exec(`
+      CREATE TABLE "${name}" (
+        ${Object.entries(schema.columns)
+          .map(([name, c]) => `"${name}" ${toSQLiteTypeName(c.type)}`)
+          .join(', ')},
+        PRIMARY KEY (${schema.primaryKey.map(k => `"${k}"`).join(', ')})
+      )`);
+
+      source = new TableSource(
+        'query.test.ts',
+        db,
+        name,
+        schema.columns,
+        schema.primaryKey,
+      );
+
+      sources.set(name, source);
+      return source;
+    },
+
+    createStorage() {
+      return new MemoryStorage();
+    },
+    addServerQuery() {
+      return () => {};
+    },
+    onTransactionCommit() {
+      return () => {};
+    },
+    batchViewUpdates<T>(applyViewUpdates: () => T): T {
+      return applyViewUpdates();
+    },
+  };
+}
